requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
chromadb>=0.5.0
